feat: generic numeric type, R-tree bulk loading, and critical performance fixes
Refactor the entire routing engine to support generic numeric weight types
(`W util.RoutingNumber`) enabling future use of int32/uint32 weights alongside
float64, and fix multiple critical performance regressions that caused p99
latency to spike from 30ms to 1s under 1500 VU load testing.

## Performance fixes

- **Restore writeJSON buffer pool**: Re-add `sync.Pool` of `bytes.Buffer` with
  streaming `sonic.NewEncoder` in `controllers/helpers.go`. The previous removal
  of this pool forced `sonic.Marshal` to allocate a fresh `[]byte` per request,
  causing GC thrashing that drained all sync.Pool objects (QueryHeaps,
  DirectionBuilders) and cascaded into 3.3GB peak memory and 1s p99. Removes
  `SetIndent` for compact JSON output.
- **Revert ScannedBitsetStorage.Clear()**: Change from `ClearAll()` (O(n)
  word-by-word zeroing) back to `bitset.New()` (O(1) pointer swap, GC handles
  old allocation). Go runtime allocates zeroed memory more efficiently via
  madvise.
- **Revert ScannedSettorage.Clear()**: Change from `Range+Remove` (O(n)
  iteration per element) back to `hashset.NewUint32WithSize()` (O(1)).
- **DirectionBuilder.done()**: Replace `clear(db.instructions)` (O(n) zeroing
  of large value-type Instruction structs) with `db.instructions[:0]` (O(1)
  reslice). Same for `turnDescriptions`.
- **DirectionBuilder.reset()**: Always reslice `[:0]` instead of conditionally
  setting to nil, preserving backing array capacity across requests.
- **Restore NewAnnotation deep copy**: Re-copy duration, distance, geometry, and
  edgeGeomOffset slices to prevent data corruption when DirectionBuilder
  reuses/reslices its backing arrays.

## R-tree spatial index overhaul

- Switch from `tidwall/rtree` to `lintang-b-s/rtree` (custom fork) with bulk
  loading via `rt.tr.Bulk()` instead of N individual Insert calls.
- Merge bidirectional road segments (same OSM way, opposite directions) into
  single R-tree leaf entries, roughly halving the number of leaves.
- Change leaf data from raw `uint64` to typed `leafData` struct with explicit
  junction flags (forward/backward head/tail).
- Replace `spatialFloor/spatialCeil` with `spatialRound` for more uniform MBR
  quantization.

## Generic type parameterization

Introduce `W util.RoutingNumber` type parameter across the core routing stack:
- `CRPRoutingEngine[W]`, `CRPALTBidirectionalSearch[W]`,
  `CRPBidirectionalSearch[W]`
- `QueryHeap[T, W]`, `DAryHeap[T, W]`, `PriorityQueueNode[T, W]`,
  `VertexInfo[W]`
- `Metric[W]`, `TimeFunction[W]`, `Landmark[W]`
- All cost functions, weight calculations, and priority comparisons now operate
  on generic `W` with `util.Lt`/`util.Gt`/`util.Eq` helpers.
- Heap ranks stored as `W` instead of `float64`, eliminating type conversion
  overhead on the hot Dijkstra path.

## Object pooling improvements

- Pool entire `CRPALTBidirectionalSearch` and `CRPBidirectionalSearch` structs
  (including embedded stalling arrays) instead of separate stalling
  entry/exit pools.
- Pool `PathUnpackerALT` instances with `DonePooled()` lifecycle.
- Pool `Coordinates` slices via `coordsPool` for path geometry assembly.
- Pool `candidatePairSet` (hashset) in RoutingService for snap pair dedup.
- Remove `routing/interface.go` (unused interface layer).

## Instruction value types

- Change `Instruction` from pointer to value type (`[]Instruction` instead of
  `[]*Instruction`), reducing per-instruction heap allocations.
- Replace `extrainfo map[string]interface{}` with typed `heading float64` and
  `hasHeading bool` fields.
- Change `RoundaboutInstruction` from pointer to embedded value type.

## Other changes

- Remove deprecated `cost_function.go` interface file.
- Remove stale test mocks (`usecases/mocks_test.go`, `usecases/routing_test.go`,
  `usecases/snap_test.go`, `usecases/mapmatcher_test.go`,
  `guidance/annotation_simplifier_test.go`).
- Add `pkg/util/binary_io.go` for binary serialization helpers.
- Add `pkg/osmparser/osm_compression.go` for OSM PBF decompression.
- Add `pkg/spatialindex/rtree_mapmatch.go` for map-matching R-tree.
- Update eval scripts and benchmarks for new generic APIs.
- Remove `JSON_BUF_POOL_SIZE` constant (32KB default used inline).

+++ eval/crp_alt/dimacs_9_shortest_path_alt/main.go
@@ -282,10 +282,10 @@ func main() {
 	}
 
 	type queryRes struct {
-		spcost float64
+		spcost int32
 	}
 
-	newQueryRes := func(spCost float64) queryRes {
+	newQueryRes := func(spCost int32) queryRes {
 		return queryRes{spcost: spCost}
 	}
